backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vswe-backend"
version = "0.1.0"
description = "Virtual Software Engineer - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "boto3>=1.35.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "python-ulid>=3.0.0",
    "PyGithub>=2.4.0",
    "PyJWT>=2.9.0",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
    "moto[dynamodb]>=5.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
